% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserRTPredictions-class.R
\name{as.data.frame,UserRTPredictions-method}
\alias{as.data.frame,UserRTPredictions-method}
\title{Convert UserRTPredictions to data.frame}
\usage{
\S4method{as.data.frame}{UserRTPredictions}(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A UserRTPredictions object}

\item{row.names}{Ignored}

\item{optional}{Ignored}

\item{...}{Ignored}
}
\value{
A data.frame of predictions
}
\description{
Convert UserRTPredictions to data.frame
}
