% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistics.R
\name{calculate_prediction_stats}
\alias{calculate_prediction_stats}
\title{Calculate Prediction Statistics}
\usage{
calculate_prediction_stats(predictions, model_results = NULL)
}
\arguments{
\item{predictions}{Output from \code{generate_all_predictions()}.}

\item{model_results}{Output from \code{build_all_models()}.}
}
\value{
A list with prediction statistics.
}
\description{
Computes statistics about predictions, if predictions have been generated.
}
