% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_cache.R
\name{compute_dataset_hash}
\alias{compute_dataset_hash}
\title{Compute Hash for Dataset RT Data}
\usage{
compute_dataset_hash(dataset)
}
\arguments{
\item{dataset}{Dataset object with rtdata component}
}
\value{
Character string with MD5 hash
}
\description{
Creates an MD5 hash of a dataset's RT data to detect changes
}
\keyword{internal}
