% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_cache.R
\name{filter_model_pairs}
\alias{filter_model_pairs}
\title{Filter Model Pairs Based on Changes}
\usage{
filter_model_pairs(model_pairs, changes, cache, export_dir)
}
\arguments{
\item{model_pairs}{List of all potential model pairs}

\item{changes}{Dataset change analysis}

\item{cache}{Build cache object}

\item{export_dir}{Directory for model outputs}
}
\value{
Filtered list of pairs that need building
}
\description{
Determines which model pairs need to be built based on dataset changes
}
\keyword{internal}
