% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserRTModels-class.R
\name{length,UserRTModels-method}
\alias{length,UserRTModels-method}
\title{Get Number of Models}
\usage{
\S4method{length}{UserRTModels}(x)
}
\arguments{
\item{x}{A UserRTModels object}
}
\value{
Integer
}
\description{
Get Number of Models
}
