% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserRTPredictions-class.R
\name{length,UserRTPredictions-method}
\alias{length,UserRTPredictions-method}
\title{Get Number of Predictions}
\usage{
\S4method{length}{UserRTPredictions}(x)
}
\arguments{
\item{x}{A UserRTPredictions object}
}
\value{
Integer
}
\description{
Get Number of Predictions
}
