% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserRTModels-class.R
\name{names,UserRTModels-method}
\alias{names,UserRTModels-method}
\title{List Available System IDs}
\usage{
\S4method{names}{UserRTModels}(x)
}
\arguments{
\item{x}{A UserRTModels object}
}
\value{
Character vector of system IDs
}
\description{
List Available System IDs
}
