% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserRTModels-class.R
\name{plot,UserRTModels,missing-method}
\alias{plot,UserRTModels,missing-method}
\title{Plot Method for UserRTModels}
\usage{
\S4method{plot}{UserRTModels,missing}(
  x,
  y,
  type = c("diagnostics", "calibration", "r_squared", "ci_widths"),
  max_plots = 9,
  ...
)
}
\arguments{
\item{x}{A UserRTModels object}

\item{y}{Ignored}

\item{type}{Plot type: "diagnostics" (default), "calibration", "r_squared", "ci_widths"}

\item{max_plots}{Maximum number of calibration plots to show (default 9)}

\item{...}{Additional arguments}
}
\value{
A ggplot object or list of ggplot objects
}
\description{
Plot Method for UserRTModels
}
