% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserRTPredictions-class.R
\name{plot,UserRTPredictions,missing-method}
\alias{plot,UserRTPredictions,missing-method}
\title{Plot Method for UserRTPredictions}
\usage{
\S4method{plot}{UserRTPredictions,missing}(x, y, type = c("ci_widths", "rt_distribution", "by_system"), ...)
}
\arguments{
\item{x}{A UserRTPredictions object}

\item{y}{Ignored}

\item{type}{Plot type: "ci_widths" (default), "rt_distribution", "by_system"}

\item{...}{Additional arguments}
}
\value{
A ggplot object
}
\description{
Plot Method for UserRTPredictions
}
