% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_cache.R
\name{purge_removed_models}
\alias{purge_removed_models}
\title{Purge Models for Removed Datasets}
\usage{
purge_removed_models(removed_ids, cache, export_dir)
}
\arguments{
\item{removed_ids}{Vector of removed dataset IDs}

\item{cache}{Build cache object}

\item{export_dir}{Directory containing model subdirectories}
}
\value{
Updated cache object
}
\description{
Deletes model directories and cache entries for models referencing removed datasets
}
\keyword{internal}
