% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{rePredRet_models}
\alias{rePredRet_models}
\title{List Available Models}
\usage{
rePredRet_models(data_path = NULL)
}
\arguments{
\item{data_path}{Path to rePredRet-data (or NULL to use cache).}
}
\value{
Tibble with model index.
}
\description{
Returns the model index with statistics for all available models.
}
\examples{
\dontrun{
models <- rePredRet_models()
}
}
