% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{rePredRet_system_predictions}
\alias{rePredRet_system_predictions}
\title{Get All Predictions for a System}
\usage{
rePredRet_system_predictions(system_id, data_path = NULL)
}
\arguments{
\item{system_id}{Target system identifier.}

\item{data_path}{Path to rePredRet-data (or NULL to use cache).}
}
\value{
Tibble with predictions.
}
\description{
Retrieves all predictions for a specific chromatographic system.
}
\examples{
\dontrun{
preds <- rePredRet_system_predictions("0001")
}
}
