% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_functions.R
\name{rePredRet_systems}
\alias{rePredRet_systems}
\title{List Available Chromatographic Systems}
\usage{
rePredRet_systems(data_path = NULL)
}
\arguments{
\item{data_path}{Path to rePredRet-data (or NULL to use cache).}
}
\value{
Tibble with system information.
}
\description{
Returns information about all chromatographic systems with available
predictions.
}
\examples{
\dontrun{
systems <- rePredRet_systems()
}
}
