% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_cache.R
\name{save_build_cache}
\alias{save_build_cache}
\title{Save Build Cache}
\usage{
save_build_cache(cache, cache_file)
}
\arguments{
\item{cache}{Cache object}

\item{cache_file}{Path to cache file}
}
\description{
Saves the build cache to disk
}
\keyword{internal}
