% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{search_compounds}
\alias{search_compounds}
\title{Batch Search for Multiple Compounds}
\usage{
search_compounds(inchis, system_type = NULL, data_path = NULL)
}
\arguments{
\item{inchis}{Character vector of InChI strings to search for.}

\item{system_type}{Optional filter by system type ("RP" or "HILIC").}

\item{data_path}{Optional path to local RepoRT data directory.}
}
\value{
A data.frame with columns: query_inchi, system_id, system_name,
system_type, compound_name, rt, inchi.
}
\description{
Searches for multiple compounds by InChI and returns all systems
where they have been measured.
}
\examples{
\dontrun{
# Search for multiple compounds
results <- search_compounds(c(
  "InChI=1S/C8H10N4O2/c1-10-4-9-6-5(10)7(13)12(3)8(14)11(6)2/h4H,1-3H3",
  "InChI=1S/C15H14O6/c16-8-4-11(18)9-6-13(20)15(21-14(9)5-8)7-1-2-10(17)12(19)3-7"
))
}

}
\seealso{
\code{\link{search_compound}} for single compound search
}
