% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserRTModels-class.R
\name{[[,UserRTModels,character-method}
\alias{[[,UserRTModels,character-method}
\title{Extract Model by System ID}
\usage{
\S4method{[[}{UserRTModels,character}(x, i)
}
\arguments{
\item{x}{A UserRTModels object}

\item{i}{System ID (character)}
}
\value{
The model object for that system
}
\description{
Extract Model by System ID
}
