% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserRTPredictions-class.R
\name{[,UserRTPredictions,ANY,ANY,ANY-method}
\alias{[,UserRTPredictions,ANY,ANY,ANY-method}
\title{Subset Predictions}
\usage{
\S4method{[}{UserRTPredictions,ANY,ANY,ANY}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{A UserRTPredictions object}

\item{i}{Row indices or logical vector}

\item{j}{Column names (optional)}

\item{drop}{Passed to data.frame subsetting}
}
\value{
Subset of predictions data.frame
}
\description{
Subset Predictions
}
