% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserRTModels-class.R
\name{summary,UserRTModels-method}
\alias{summary,UserRTModels-method}
\title{Summary Method for UserRTModels}
\usage{
\S4method{summary}{UserRTModels}(object, ...)
}
\arguments{
\item{object}{A UserRTModels object}
}
\value{
A list with summary statistics
}
\description{
Summary Method for UserRTModels
}
