% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserRTPredictions-class.R
\name{summary,UserRTPredictions-method}
\alias{summary,UserRTPredictions-method}
\title{Summary Method for UserRTPredictions}
\usage{
\S4method{summary}{UserRTPredictions}(object, ...)
}
\arguments{
\item{object}{A UserRTPredictions object}
}
\value{
A list with summary statistics
}
\description{
Summary Method for UserRTPredictions
}
