% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_cache.R
\name{update_dataset_cache}
\alias{update_dataset_cache}
\title{Update Dataset Cache}
\usage{
update_dataset_cache(datasets, cache, new_ids, changed_ids)
}
\arguments{
\item{datasets}{Named list of datasets}

\item{cache}{Build cache object}

\item{new_ids}{Vector of new dataset IDs}

\item{changed_ids}{Vector of changed dataset IDs}
}
\value{
Updated cache object
}
\description{
Updates cache with current dataset hashes
}
